backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==41.0.7
cssselect2==0.7.0
distlib==0.3.8
Django==5.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.12.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-session-timeout==0.1.0
django-simple-history==3.4.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-flex-fields==1.0.2
drf-writable-nested==0.7.0
dulwich==0.21.7
fastjsonschema==2.19.0
filelock==3.13.1
freetype-py==2.3.0
html5lib==1.1
idna==3.6
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
lxml==4.9.3
more-itertools==10.1.0
msgpack==1.0.7
num2words==0.5.13
oscrypto==1.3.0
packaging==23.2
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.2
PyJWT==2.8.0
pypdf==3.17.2
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-bidi==0.4.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.5.2
reportlab==4.0.8
requests==2.31.0
requests-toolbelt==1.0.0
rlPyCairo==0.3.0
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomlkit==0.12.3
trove-classifiers==2023.11.29
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.25.0
webencodings==0.5.1
wheel==0.42.0
xhtml2pdf==0.2.13
xlwt==1.3.0
